+++ Villeon/Villeon/Systems/Update/PlayerTopDownMovementSystem.cs
@@ -25,15 +25,15 @@ namespace Villeon.Systems
 
         public Signature Signature { get; private set; } = new();
 
-        public void Update(double time)
+        public void Update(float time)
         {
             float leftRightAxis = KeyHandler.IsPressed(Keys.A) ? -1 : KeyHandler.IsPressed(Keys.D) ? 1 : 0;
             float topDownAxis = KeyHandler.IsPressed(Keys.S) ? -1 : KeyHandler.IsPressed(Keys.W) ? 1 : 0;
-            
+
             foreach (IEntity entity in Entities)
             {
                 Collider playerCollider = entity.GetComponent<Collider>();
-                playerCollider.Position += new Vector2(leftRightAxis, topDownAxis) * (float)time * Constants.TOPDOWNMOVEMENTSPEED;
+                playerCollider.Position += new Vector2(leftRightAxis, topDownAxis) * time * Constants.TOPDOWNMOVEMENTSPEED;
             }
         }
     }
